Move parsing of PSMs to separate module
