+++ ms2rescore/report/generate.py
@@ -48,6 +48,7 @@ def generate_report(
     output_path_prefix: str,
     psm_list: Optional[psm_utils.PSMList] = None,
     feature_names: Optional[Dict[str, list]] = None,
+    use_txt_log: bool = False,
 ):
     """
     Generate the report.
@@ -63,9 +64,12 @@ def generate_report(
     feature_names
         Feature names to be used for the report. If not provided, the feature names will be
         read from the feature names file that matches the ``output_path_prefix``.
+    use_txt_log
+        If True, the log file will be read from ``output_path_prefix + ".log.txt"`` instead of
+        ``output_path_prefix + ".log.html"``.
 
     """
-    files = _collect_files(output_path_prefix)
+    files = _collect_files(output_path_prefix, use_txt_log=use_txt_log)
 
     # Read PSMs
     if not psm_list:
@@ -131,7 +135,7 @@ def generate_report(
     _render_and_write(output_path_prefix, **context)
 
 
-def _collect_files(output_path_prefix):
+def _collect_files(output_path_prefix, use_txt_log=False):
     """Collect all files generated by MS²Rescore."""
     logger.info("Collecting files...")
     files = {
@@ -139,9 +143,9 @@ def _collect_files(output_path_prefix):
         "configuration": Path(output_path_prefix + ".full-config.json").resolve(),
         "feature names": Path(output_path_prefix + ".feature_names.tsv").resolve(),
         "feature weights": Path(output_path_prefix + ".mokapot.weights.tsv").resolve(),
-        "log": Path(output_path_prefix + ".log.html").resolve()
-        if Path(output_path_prefix + ".log.html").is_file()
-        else Path(output_path_prefix + ".log.txt").resolve(),
+        "log": Path(output_path_prefix + ".log.txt").resolve()
+        if use_txt_log
+        else Path(output_path_prefix + ".log.html").resolve(),
     }
     for file, path in files.items():
         if Path(path).is_file():
